requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
anyjson==0.3.3
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
cryptography==2.9.2
execnet==1.7.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
kombu==3.0.37
lxml==4.2.5
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.4
paramiko==2.7.1
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pycryptodome==3.9.7
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==5.4.3
pytest-assume==2.2.1
pytest-forked==1.1.3
pytest-repeat==0.7.0
pytest-rerunfailures==9.0
pytest-xdist==1.23.2
pytz==2020.1
PyYAML==3.12
requests==2.23.0
requests-toolbelt==0.9.1
selenium==2.53.6
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
